[1194] Amended to use an onboarding controller
