+++ app/controllers/support/providers_controller.rb
@@ -17,10 +17,6 @@ module Support
       @provider.organisations.build
     end
 
-    def new_provider
-      @provider = ProviderForm.new(current_user, recruitment_cycle:)
-    end
-
     def edit
       provider
     end
@@ -31,16 +27,6 @@ module Support
         redirect_to support_recruitment_cycle_provider_path(provider.recruitment_cycle_year, provider), flash: { success: 'Provider was successfully created' }
       else
         render :new
-      end
-    end
-
-    def create_provider
-      @provider = ProviderForm.new(current_user, recruitment_cycle:, params: provider_form_params)
-
-      if @provider.valid?
-        redirect_to support_recruitment_cycle_providers_path(recruitment_cycle.year), flash: { success: 'Provider was successfully created' }
-      else
-        render :new_provider
       end
     end
 
@@ -97,11 +83,6 @@ module Support
                                                                         address4
                                                                         postcode],
                                                    organisations_attributes: %i[name]).merge(recruitment_cycle:)
-    end
-
-    def provider_form_params
-      params.require(:support_provider_form)
-            .permit(ProviderForm::FIELDS)
     end
   end
 end
