+++ config/routes/support.rb
@@ -4,9 +4,10 @@ namespace :support do
   get '/', to: 'recruitment_cycle#index'
 
   resources :recruitment_cycles, param: :year, constraints: { year: /#{Settings.current_recruitment_cycle_year}|#{Settings.current_recruitment_cycle_year + 1}/ }, path: '' do
+    namespace :providers do
+      resource :onboarding, only: %i[new create]
+    end
     resources :providers, except: %i[destroy] do
-      get '/new-provider', on: :collection, to: 'providers#new_provider'
-      post '/new-provider', on: :collection, to: 'providers#create_provider'
       resource :check_user, only: %i[show update], controller: 'providers/users_check', path: 'users/check'
       resources :users, controller: 'providers/users' do
         member do
